refactor optuna
