+++ .gitignore
@@ -1,4 +1,5 @@
 lightning_logs/
+optuna.db
 
 #   Byte-compiled / optimized / DLL files
 __pycache__/
