+++ examples/robot_localization_failure/config.yaml
@@ -36,3 +36,5 @@ learning:
   learning_rate: 0.0001
   epochs: 2
   model_path: "models/robot_localization.model"
+optuna:
+  storage: "sqlite:///optuna.db"
