+++ examples/robot_localization_failure/pyproject.toml
@@ -12,6 +12,7 @@ dependencies = [
   "plotly>=6.0.1",
   "dash>=3.0.4",
   "optuna>=4.5.0",
+  "polars-u64-idx>=1.32.3",
 ]
 
 [tool.pyright]
