+++ examples/robot_localization_failure/run_optuna.py
@@ -1,319 +1,83 @@
 #!/usr/bin/env python
+
 import logging
-from os import PathLike
-from pathlib import Path
-import optuna
 from functools import partial
+from pathlib import Path
+
+import optuna
 import polars as pl
-from architectures.cnn import CNN
-from custom_learners.image_based_lightning_learner import (
-    ImageBasedLightningLearner,
-    ImageBasedPyTorchModel,
-)
-from custom_transforms.collapse import Collapse
-from custom_transforms.detect_delocalizations import DetectDelocalizations
-from custom_transforms.localization_status import LocalizationStatus
-from custom_transforms.slice_time_series import SliceTimeSeries
-from custom_transforms.zero_order_hold_matching import ZeroOrderHold
 from omegaconf import DictConfig, ListConfig
+from optuna import Trial
+from training import (
+    collect_data,
+    evaluate,
+    train,
+)
 
 import flowcean.cli
-from flowcean.core import Transform
-from flowcean.core.strategies import evaluate_offline
-from flowcean.core.transform import Lambda
-from flowcean.polars import DataFrame
-from flowcean.polars.transforms.drop import Drop
-from flowcean.ros import load_rosbag
-from flowcean.sklearn.metrics.classification import (
-    Accuracy,
-    ClassificationReport,
-    FBetaScore,
-    PrecisionScore,
-    Recall,
-)
 
 logger = logging.getLogger(__name__)
 
 
-def define_transforms(
-    position_threshold: float,
-    heading_threshold: float,
-) -> Transform:
-    def convert_map_to_bool(df: pl.LazyFrame) -> pl.LazyFrame:
-        return df.with_columns(
-            pl.col("/map").struct.with_fields(
-                pl.field("data").list.eval(pl.element() != 0),
-            ),
-        )
-
-    return (
-        Collapse("/map", element=0)
-        | Lambda(convert_map_to_bool)
-        | ZeroOrderHold(
-            features=[
-                "/scan",
-                "/particle_cloud",
-                "/momo/pose",
-                "/amcl_pose",
-            ],
-            name="measurements",
-        )
-        | Drop("/scan", "/particle_cloud", "/momo/pose", "/amcl_pose")
-        | DetectDelocalizations("/delocalizations", name="slice_points")
-        | Drop("/delocalizations")
-        | SliceTimeSeries(
-            time_series="measurements",
-            slice_points="slice_points",
-        )
-        | Drop("slice_points")
-        | LocalizationStatus(
-            time_series="measurements",
-            ground_truth="/momo/pose",
-            estimation="/amcl_pose",
-            position_threshold=position_threshold,
-            heading_threshold=heading_threshold,
-        )
-    )
-
-
-def load_and_process_rosbag(
-    path: str | PathLike,
-    config: DictConfig | ListConfig,
-) -> pl.DataFrame:
-    cache_path = Path(path).with_suffix(".processed.parquet")
-    if cache_path.exists():
-        logger.info(
-            "Loading already processed rosbag from cache: %s",
-            cache_path,
-        )
-        return pl.read_parquet(cache_path)
-
-    logger.info("Processing rosbag: %s", path)
-    data = load_rosbag(
-        path=path,
-        topics={
-            "/amcl_pose": [
-                "pose.pose.position.x",
-                "pose.pose.position.y",
-                "pose.pose.orientation.x",
-                "pose.pose.orientation.y",
-                "pose.pose.orientation.z",
-                "pose.pose.orientation.w",
-            ],
-            "/momo/pose": [
-                "pose.position.x",
-                "pose.position.y",
-                "pose.orientation.x",
-                "pose.orientation.y",
-                "pose.orientation.z",
-                "pose.orientation.w",
-            ],
-            "/scan": [
-                "ranges",
-                "angle_min",
-                "angle_max",
-                "angle_increment",
-                "range_min",
-                "range_max",
-            ],
-            "/map": [
-                "data",
-                "info.resolution",
-                "info.width",
-                "info.height",
-                "info.origin.position.x",
-                "info.origin.position.y",
-                "info.origin.position.z",
-                "info.origin.orientation.x",
-                "info.origin.orientation.y",
-                "info.origin.orientation.z",
-                "info.origin.orientation.w",
-            ],
-            "/delocalizations": ["data"],
-            "/particle_cloud": ["particles"],
-        },
-        message_paths=config.rosbag.message_paths,
-    )
-    transform = define_transforms(
-        config.localization.position_threshold,
-        config.localization.heading_threshold,
-    )
-    transformed_data: pl.DataFrame = transform.apply(data).collect(
-        engine="streaming",
-    )
-
-    logger.info("Caching processed data to Parquet file: %s", cache_path)
-    transformed_data.write_parquet(cache_path)
-
-    return transformed_data
-
-
-def explode_and_collect_samples(data: pl.DataFrame) -> pl.DataFrame:
-    return (
-        data.explode("measurements")
-        .unnest("measurements")
-        .unnest("value")
-        .select(
-            pl.col("/map"),
-            pl.struct(
-                [
-                    pl.col("/scan/ranges").alias("ranges"),
-                    pl.col("/scan/angle_min").alias("angle_min"),
-                    pl.col("/scan/angle_max").alias("angle_max"),
-                    pl.col("/scan/angle_increment").alias(
-                        "angle_increment",
-                    ),
-                    pl.col("/scan/range_min").alias("range_min"),
-                    pl.col("/scan/range_max").alias("range_max"),
-                ],
-            ).alias("/scan"),
-            pl.col("/particle_cloud/particles").alias(
-                "/particle_cloud",
-            ),
-            pl.struct(
-                [
-                    pl.struct(
-                        [
-                            pl.col(
-                                "/amcl_pose/pose.pose.position.x",
-                            ).alias("position.x"),
-                            pl.col(
-                                "/amcl_pose/pose.pose.position.y",
-                            ).alias("position.y"),
-                            pl.col(
-                                "/amcl_pose/pose.pose.orientation.x",
-                            ).alias("orientation.x"),
-                            pl.col(
-                                "/amcl_pose/pose.pose.orientation.y",
-                            ).alias("orientation.y"),
-                            pl.col(
-                                "/amcl_pose/pose.pose.orientation.z",
-                            ).alias("orientation.z"),
-                            pl.col(
-                                "/amcl_pose/pose.pose.orientation.w",
-                            ).alias("orientation.w"),
-                        ],
-                    ).alias("pose"),
-                ],
-            ).alias("/amcl_pose"),
-            pl.col("is_delocalized"),
-        )
-    )
-
-
-def train_and_evaluate(
+def objective(
+    trial: Trial,
     train_data: pl.DataFrame,
-    test_data: pl.DataFrame,
+    eval_data: pl.DataFrame,
     config: DictConfig | ListConfig,
-) -> ImageBasedPyTorchModel:
-    learner = ImageBasedLightningLearner(
-        module=CNN(
-            image_size=config.architecture.image_size,
-            in_channels=3,
-            learning_rate=config.learning.learning_rate,
-        ),
-        batch_size=config.learning.batch_size,
-        max_epochs=config.learning.epochs,
-        image_size=config.architecture.image_size,
-        width_meters=config.architecture.width_meters,
-    )
-    logger.info("Training model for %s epochs", config.learning.epochs)
-    model = learner.learn(
-        inputs=train_data.drop(["is_delocalized"]),
-        outputs=train_data.select(["is_delocalized"]),
-    )
-
-    # # Evaluate the model
-    # metrics = [
-    #     Accuracy(),
-    #     ClassificationReport(),
-    #     FBetaScore(beta=1.0),
-    #     PrecisionScore(),
-    #     Recall(),
-    # ]
-    # logger.info("Evaluating model on test data")
-    # report = evaluate_offline(
-    #     model,
-    #     DataFrame(test_data),
-    #     inputs=["/map", "/scan", "/particle_cloud", "/amcl_pose"],
-    #     outputs=["is_delocalized"],
-    #     metrics=metrics,
-    # )
-    # print(report)
-    return model
-
-def objective(trial, train_data, eval_data, config):
+) -> float:
     config.learning.learning_rate = trial.suggest_float(
-        "learning_rate", 1e-5, 1e-2, log=True
+        "learning_rate",
+        low=1e-5,
+        high=1e-2,
+        log=True,
     )
     config.learning.batch_size = trial.suggest_categorical(
-        "batch_size", [32, 64, 128, 256]
+        "batch_size",
+        [32, 64, 128, 256],
     )
     config.architecture.image_size = trial.suggest_categorical(
-        "image_size", [64, 128, 150, 224]
+        "image_size",
+        [64, 128, 150, 224],
     )
     config.architecture.width_meters = trial.suggest_float(
-        "width_meters", 5.0, 50.0
+        "width_meters",
+        5.0,
+        50.0,
     )
 
-    model = train_and_evaluate(
+    model = train(
         train_data=train_data,
-        test_data=eval_data,
         config=config,
     )
-
-    metrics = [ 
-        Accuracy(),
-        ClassificationReport(),
-        FBetaScore(beta=1.0),
-        PrecisionScore(),
-        Recall(),
-    ]
-    report = evaluate_offline(
-        model,
-        DataFrame(eval_data),
-        inputs=["/map", "/scan", "/particle_cloud", "/amcl_pose"],
-        outputs=["is_delocalized"],
-        metrics=metrics,
+    report = evaluate(
+        model=model,
+        test_data=eval_data,
     )
+
     f1_score = report["FBetaScore"]
     logger.info("Trial finished with F1=%.4f", f1_score)
 
-    return f1_score
+    return f1_score  # pyright: ignore[reportReturnType]
 
 
 def main() -> None:
     config = flowcean.cli.initialize()
 
-    logger.info("Collecting training data")
-    runs_train = [
-        load_and_process_rosbag(
-            path=path,
-            config=config,
-        )
-        for path in config.rosbag.training_paths
-    ]
-    logger.info("Combining training data")
-    samples_train = explode_and_collect_samples(
-        pl.concat(runs_train, how="vertical"),
-    )
+    samples_train, samples_eval = collect_data(config)
 
-    logger.info("Collecting evaluation data")
-    runs_eval = [
-        load_and_process_rosbag(
-            path=path,
-            config=config,
-        )
-        for path in config.rosbag.evaluation_paths
-    ]
-    logger.info("Combining evaluation data")
-    samples_eval = explode_and_collect_samples(
-        pl.concat(runs_eval, how="vertical"),
-    )
     logger.info("Starting Optuna optimization")
-    study = optuna.create_study(direction="maximize")
+    study = optuna.create_study(
+        direction="maximize",
+        storage=config.optuna.storage,
+        sampler=optuna.samplers.TPESampler(),
+    )
     study.optimize(
-        partial(objective, train_data=samples_train, eval_data=samples_eval, config=config),
+        partial(
+            objective,
+            train_data=samples_train,
+            eval_data=samples_eval,
+            config=config,
+        ),
         n_trials=2,
     )
 
@@ -325,12 +89,11 @@ def main() -> None:
     config.learning.batch_size = best_params["batch_size"]
     config.architecture.image_size = best_params["image_size"]
     config.architecture.width_meters = best_params["width_meters"]
-    
+
     logger.info("Retraining final model with best parameters")
 
-    final_model = train_and_evaluate(
+    final_model = train(
         train_data=samples_train,
-        test_data=samples_eval,
         config=config,
     )
